Fix for handling manually suppressing error modals
