+++ src/ErrorHandling/ErrorState.tsx
@@ -1,9 +1,4 @@
 import { EventEmitter } from "../utils/EventEmitter";
-import { Settings } from "../Settings/Settings";
-import { Router } from "../ui/GameRoot";
-import { SimplePage } from "@enums";
-
-let currentId = 0;
 
 export type ErrorRecord = {
   id: number;
@@ -23,84 +18,18 @@ export const ErrorState = {
   ActiveError: null as ErrorRecord | null,
   Errors: [] as ErrorRecord[],
   UnreadErrors: 0,
-  PreventModalsUntil: Settings.SuppressErrorModals ? new Date("3000-01-01") : new Date(),
+  PreventModalsUntil: new Date(),
 };
 
 export function errorModalsAreSuppressed(): boolean {
   return ErrorState.PreventModalsUntil > new Date();
 }
 
-export function toggleSuppressErrorModals(newValue: boolean): void {
+export function toggleSuppressErrorModals(newValue: boolean, indefinite = false): void {
   if (newValue) {
-    ErrorState.PreventModalsUntil = new Date(Date.now() + 1000 * 60 * 5); // Suppress for 5 minutes
+    ErrorState.PreventModalsUntil = new Date(indefinite ? new Date("3000-01-01") : Date.now() + 1000 * 60 * 5);
+    ErrorState.ActiveError = null;
   } else {
     ErrorState.PreventModalsUntil = new Date();
-  }
-}
-
-export function toggleSuppressErrorModalsSetting(newValue: boolean): void {
-  Settings.SuppressErrorModals = newValue;
-  if (newValue) {
-    ErrorState.PreventModalsUntil = new Date("3000-01-01");
-  } else {
-    ErrorState.PreventModalsUntil = new Date();
-  }
-}
-
-export const DisplayError = (
-  message: string,
-  errorType: string,
-  scriptName = "",
-  hostname: string = "",
-  pid: number = -1,
-) => {
-  const errorPageOpen = Router.page() === SimplePage.RecentErrors;
-  if (!errorPageOpen) {
-    ErrorState.UnreadErrors++;
-  }
-  const prior = findExistingErrorCopy(message, hostname);
-  if (prior) {
-    prior.occurrences++;
-    prior.time = new Date();
-    if (pid !== -1) {
-      prior.pid = pid;
-    }
-    prior.server = hostname;
-    prior.message = message;
-
-    updateActiveError(prior);
-  } else {
-    ErrorState.Errors.unshift({
-      id: currentId++,
-      server: hostname,
-      errorType,
-      scriptName,
-      message,
-      pid,
-      occurrences: 1,
-      time: new Date(),
-      unread: !errorPageOpen,
-    });
-    while (ErrorState.Errors.length > Settings.MaxRecentScriptsCapacity) {
-      ErrorState.Errors.pop();
-    }
-    updateActiveError(ErrorState.Errors[0]);
-  }
-};
-
-function findExistingErrorCopy(message: string, hostname: string): ErrorRecord | null {
-  const serverAgnosticMessage = message.replaceAll(hostname, "<server>");
-  return (
-    ErrorState.Errors.find(
-      (e) => e.message.replaceAll(e.server, "<server>") === serverAgnosticMessage || e.message === message,
-    ) ?? null
-  );
-}
-
-function updateActiveError(error: ErrorRecord): void {
-  // TODO: Fix this bug
-  if (!ErrorState.ActiveError) {
-    ErrorState.ActiveError = error;
-    ErrorState.ErrorUpdate.emit(ErrorState.ActiveError);
   }
 }
