+++ src/Settings/Settings.ts
@@ -12,6 +12,7 @@ import {
   assertAndSanitizeStyles,
 } from "../JsonSchema/JSONSchemaAssertion";
 import { mergePlayerDefinedKeyBindings, type PlayerDefinedKeyBindingsType } from "../utils/KeyBindingUtils";
+import { toggleSuppressErrorModals } from "../ErrorHandling/ErrorState";
 
 /**
  * This function won't be able to catch **all** invalid hostnames. In order to validate a hostname properly, we need to
@@ -255,5 +256,8 @@ export const Settings = {
 
     // Merge Settings.KeyBindings with DefaultKeyBindings.
     mergePlayerDefinedKeyBindings(Settings.KeyBindings);
+
+    // Set up initial state for error modal suppression
+    toggleSuppressErrorModals(Settings.SuppressErrorModals, true);
   },
 };
